Address PR comments and update tests
