+++ tests/test_pfar.py
@@ -1,17 +1,16 @@
 import json
 import torch as th
 import sys
-sys.path.append('/home/XLL1713/POPROX/poprox/poprox-recommender/src')
-print(sys.path)
+# sys.path.append('PATH_TO_src_dir'), should not need this when model path func is ready
 from poprox_concepts import Article, ClickHistory
-from poprox_recommender.handler import general_topics, match_news_topics_to_general, extract_general_topic
+from poprox_recommender.handler import extract_general_topic
+from poprox_recommender.topics import general_topics, match_news_topics_to_general
 from poprox_recommender.default import select_articles, user_topic_preference
-from poprox_recommender.paths import project_root, model_file_path
 from safetensors.torch import load_file
 
 
 def load_test_articles():
-    event_path = "/home/XLL1713/POPROX/poprox/poprox-recommender/tests/request_body.json" # might need to config the path correctly
+    event_path = "request_body.json" # update when the model path func is ready
     with open(event_path, 'r') as j:
         req_body = json.loads(j.read())
 
@@ -25,8 +24,9 @@ def load_test_articles():
 
     return todays_articles, past_articles, click_history, num_recs
 
+
 def test_topic_classification():
-    todays_articles = load_test_articles()
+    todays_articles, _, _, _ = load_test_articles()
     topic_matched_dict, todays_article_matched_topics = match_news_topics_to_general(todays_articles)
     print('***************** topic matched dict *****************')
     print(topic_matched_dict)
@@ -36,18 +36,19 @@ def test_topic_classification():
 
 
 def test_extract_generalized_topic():
-    todays_articles = load_test_articles()
+    todays_articles, _, _, _ = load_test_articles()
     for article in todays_articles:
         generalized_topics = extract_general_topic(article)
         for topic in generalized_topics:
             assert topic in general_topics
+
 
 def load_model(device_name=None):
 
     if device_name is None:
         device_name = "cuda" if th.cuda.is_available() else "cpu"
 
-    load_path = model_file_path("model.safetensors")
+    load_path = "model.safetensors" # update when the model path func is ready
     checkpoint = load_file(load_path)
 
     return checkpoint, device_name
@@ -56,17 +57,13 @@ def load_model(device_name=None):
 if __name__ == '__main__':
     todays_articles, past_articles, click_data, num_recs = load_test_articles()
 
-    MODEL, DEVICE = load_model()
-    TOKEN_MAPPING = 'distilbert-base-uncased'
     user_preference_dict = user_topic_preference(past_articles, click_data)
     algo_params = {'user_topic_preference': user_preference_dict}
     recommendations = select_articles(
         todays_articles,
         past_articles,
         click_data,
-        MODEL,
-        DEVICE,
-        TOKEN_MAPPING,
         num_recs,
         algo_params
     )
+    print(recommendations)
